+++ submitr/tests/test_submission.py
@@ -1373,7 +1373,7 @@ class Scenario:
 
 @mock.patch.object(submission_module, "_get_health_page")
 @mock.patch.object(submission_module, "DEBUG_PROTOCOL", False)
-def test_submit_any_ingestion_old_protocol(mock_get_health_page):
+def obsolete_test_submit_any_ingestion_old_protocol(mock_get_health_page):
 
     mock_get_health_page.return_value = {HealthPageKey.S3_ENCRYPT_KEY_ID: TEST_ENCRYPT_KEY}